third_party/WebKit/LayoutTests/inspector/sass/test-ast-css-3.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="./sass-test.js"></script>
<style>
pre {
    font-family: monospace;
}
</style>
<script>

function test()
{
    InspectorTest.evaluateInPage("getCSS()", onCSS);

    function onCSS(result)
    {
        InspectorTest.parseCSS("", result.value)
            .then(InspectorTest.dumpAST)
            .then(InspectorTest.validateASTRanges)
            .catch(console.error.bind(console))
            .then(InspectorTest.completeTest);
    }
}

function getCSS()
{
    return document.querySelector(".snippet").textContent;
}

</script>
</head>

<body onload="runTest()">
<p>
Verifies AST of CSS with missing trailing semicolon after last property.
</p>
<pre class="snippet">
body {
    /* padding-top: 5em; */
    margin: 0 auto
}

body {
    color: red;
    /* margin: 0 auto */
}

</pre>
</body>
</html>
