third_party/WebKit/LayoutTests/inspector/sass/test-ast-css-3-expected.txt

Verifies AST of CSS with missing trailing semicolon after last property.

body {
    /* padding-top: 5em; */
    margin: 0 auto
}

body {
    color: red;
    /* margin: 0 auto */
}

=== AST === 
rule 0: "body"
    property 0
        name: "padding-top"
            range: {"startLine":1,"startColumn":7,"endLine":1,"endColumn":18}
        value: "5em"
            range: {"startLine":1,"startColumn":20,"endLine":1,"endColumn":23}
        range: {"startLine":1,"startColumn":4,"endLine":1,"endColumn":27}
        disabled: true
    property 1
        name: "margin"
            range: {"startLine":2,"startColumn":4,"endLine":2,"endColumn":10}
        value: "0 auto"
            range: {"startLine":2,"startColumn":12,"endLine":2,"endColumn":18}
        range: {"startLine":2,"startColumn":4,"endLine":3,"endColumn":0}
        disabled: false
rule 1: "body"
    property 0
        name: "color"
            range: {"startLine":6,"startColumn":4,"endLine":6,"endColumn":9}
        value: "red"
            range: {"startLine":6,"startColumn":11,"endLine":6,"endColumn":14}
        range: {"startLine":6,"startColumn":4,"endLine":6,"endColumn":15}
        disabled: false
    property 1
        name: "margin"
            range: {"startLine":7,"startColumn":7,"endLine":7,"endColumn":13}
        value: "0 auto"
            range: {"startLine":7,"startColumn":15,"endLine":7,"endColumn":21}
        range: {"startLine":7,"startColumn":4,"endLine":7,"endColumn":24}
        disabled: true
======
Ranges OK.
